moves service definition to main
